docs(forms): add better documentation for custom form fields
affects: @medly-components/forms
