+++ packages/forms/src/components/Fields/Fields.components.tsx
@@ -33,7 +33,7 @@ export const Fields: React.FC<FieldsProps> = React.memo(props => {
             <>
                 {Object.keys(fields).map(key => {
                     // @ts-ignore
-                    const { gridColumn, gridRow, Component, ...componentProps } = fields[key],
+                    const { gridColumn, gridRow, component: Component, ...componentProps } = fields[key],
                         // @ts-ignore
                         disabled = props.disabled || componentProps.disabled,
                         showDecorators = props.showDecorators,
